refactor the schema  folder add intenral file
